requirements.txt
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mysql-connector==2.1.4
simplejson==3.13.2
six==1.11.0
Werkzeug==0.12.2
